Adds new Company page
