--- front/src/company/Company.scss
@@ -1,43 +1,0 @@
-p {
-  line-height: 1em;
-}
-
-.box {
-  width: 100%;
-  border-radius: 2px;
-  border: 1px solid #c9d3df;
-  padding: 10px;
-  min-width: 0;
-}
-
-.columns {
-  display: flex;
-  justify-content: space-around;
-  margin: 0 0 10px 0;
-}
-
-.section {
-  padding: 2em 0;
-}
-
-.company__info {
-}
-
-.company__logo {
-  width: 150px;
-  height: auto;
-}
-
-table {
-  margin-top: 20px;
-}
-
-.table__container {
-  max-height: 600px;
-  overflow: scroll;
-}
-
-.table__filter {
-  margin-top: 1em;
-  max-width: 20em;
-}
