--- front/src/company/CompanyContact.scss
@@ -1,15 +1,0 @@
-.company__item {
-  margin-bottom: 1em;
-}
-
-.company__item-key {
-  display: block;
-  color: #53657d;
-  margin-bottom: 0;
-}
-
-.url-ellipsis {
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-}
