--- front/src/company/CompanyMap.scss
@@ -1,6 +1,0 @@
-#map {
-  padding: 0;
-  height: 300px;
-  flex: 0 0 40%;
-  margin-left: 10px;
-}
